LargeLocal

+++ src/game/Com/Com.cpp
@@ -188,7 +188,7 @@ int Com_sprintf(char* buf, size_t bufCount, char* fmt, ...)
 
 
 //DONE : 0x416E40
-void* Com_Memcpy(void* dest, const void* src, size_t size)
+void* Com_Memcpy(void* dest, const void* src, int size)
 {
     return memcpy(dest, src, size); //seems like Com_Memcpy appears to be a wrapper across a few engines
 }
